requirements.txt
bokeh~=2.0
bs4~=0.0
Django==3.*
djangorestframework~=3.11
django-extensions~=2.2
django-filter~=2.2
nipype~=1.5
numpy~=1.19
tqdm~=4.40
niwidgets~=0.2
git+https://github.com/TheLabbingProject/django_analyses.git@master#egg=django_analyses
git+https://github.com/TheLabbingProject/django_dicom.git@master#egg=django_dicom
git+https://github.com/ZviBaratz/dicom_parser.git@master#egg=dicom_parser